pyproject.toml
[project]
name = "fitlog"
version = "0.1.0"
description = "Google Fit health data tracking system with InfluxDB and Grafana"
authors = [
    { name = "orangekame3", email = "orangekame3@example.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
    "google-auth>=2.23.4",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.108.0",
    "influxdb-client>=1.38.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pytz>=2023.3",
    "click>=8.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-watch>=4.2.0",
    "ruff>=0.1.0",
    "ty>=0.0.1a14",
]

[project.scripts]
fitlog-fetch = "fitlog.fetch:main"
fitlog-influx-test = "fitlog.influx_writer:main"
fitlog-mock = "fitlog.mock_data:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-watch>=4.2.0",
    "ruff>=0.1.0",
    "ty>=0.0.1a14",
]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "Q",  # flake8-quotes
]
ignore = [
    "E501",  # line too long, handled by ruff format
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["B011"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["fitlog"]

[tool.ty]
[tool.ty.environment]
root = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=fitlog --cov-report=html --cov-report=term-missing --tb=short"
minversion = "6.0"

[tool.hatch.build.targets.wheel]
packages = ["fitlog"]